fixed mapping from Numeric to Float

+++ hotelapp/src/main/java/pl/mgarbowski/hotelapp/domain/apartment/Apartment.java
@@ -26,5 +26,5 @@ public class Apartment {
     private Hotel hotel;
     @ManyToOne(optional = false)
     private Currency currency;
-    private Float avg_rating;
+    private BigDecimal avg_rating;
 }
