+++ hotelapp/src/main/java/pl/mgarbowski/hotelapp/domain/hotel/Hotel.java
@@ -6,6 +6,8 @@ import jakarta.persistence.OneToOne;
 import lombok.Data;
 import pl.mgarbowski.hotelapp.domain.address.Address;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 public class Hotel {
@@ -16,5 +18,5 @@ public class Hotel {
     private Integer stars;
     @OneToOne
     private Address address;
-    private Float avg_rating;
+    private BigDecimal avg_rating;
 }
